+++ provision/install.sh
@@ -9,9 +9,9 @@ apt-get update
 apt-get upgrade -y
 
 # Install common packages.
-apt-get install -y --no-install-recommends software-properties-common \
-build-essential curl dos2unix gcc libmcrypt4 libpcre3-dev make \
-graphicsmagick sqlite3 libsqlite3-dev memcached ssl-cert ntp
+apt-get install -y --no-install-recommends \
+software-properties-common build-essential curl dos2unix gcc libmcrypt4 libpcre3-dev make ntp \
+apt-transport-https ca-certificates graphicsmagick sqlite3 libsqlite3-dev memcached ssl-cert
 
 # Install services.
 source /vagrant/provision/installation/git.sh
@@ -20,6 +20,7 @@ source /vagrant/provision/installation/mysql.sh
 source /vagrant/provision/installation/nginx.sh
 source /vagrant/provision/installation/apache.sh
 source /vagrant/provision/installation/php.sh
+source /vagrant/provision/installation/docker.sh
 source /vagrant/provision/installation/postfix.sh
 source /vagrant/provision/installation/dovecot.sh
 source /vagrant/provision/installation/node.sh